feat: DataPatternDetector模式检测器+DataStreamRecorder数据流记录器+多模块增强
新增DataPatternDetector(84+138=222行): 滑动窗口+哈希表统计发现重复字节序列，
可配置最小/最大模式长度(2-16)和最小出现次数(2)，按频率降序排序，模式位置追踪。
新增DataStreamRecorder(276+561=837行): 原始数据流捕获+标注+自动分割+3种导出格式。
ModbusMaster重构(195+294行): 协议方法签名统一+增强的RTU通信+统计查询。
PacketVisualizer/ProtocolSession/WaveformFilter/SerialFrameDecoder/
DataEncryptionEngine/SerialScriptEngine/SendHistory等20+模块编译修复和增强。

评分: 336 + 1 = 337
变更: 30 files, +6149 insertions, -239 deletions

+++ src/serial/commands/SendHistoryStats.cpp
@@ -63,7 +63,7 @@ QString SendHistory::statisticsSummary() const
 }
 
 /** @brief 获取历史总发送次数 @return 总发送次数（含去重） */
-int SendHistory::totalSendCount() const
+quint64 SendHistory::totalSendCount() const
 {
     return m_totalSendCount;
 }
